+++ app/core/config.py
@@ -13,6 +13,9 @@ class Settings(BaseSettings):
     PROJECT_NAME: str = "aviauth-api"
     VERSION: str = "0.1.0"
 
+    MAX_DB_CONNECTION_RETRIES = 5 * 60
+    DB_CONNECTION_RETRY_WAIT_SECONDS = 5
+
     DEBUG: bool = os.getenv("DEBUG", False)
 
     @validator("SQLALCHEMY_DATABASE_URL", pre=True)