+++ start.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
-# Use this script to start the application with hot reload.
 
+# Start the application with hot reload.
 uvicorn app.main:app --port 8000 --reload --log-level info